fix(dev-mode): apply ce-review safe + approved fixes
Findings from the dev-mode /ce-review (10 reviewers). Applied:

- dev_telemetry: clamp MB ceiling to INT32_MAX so totals stay safe to
  cast to the int32 LVGL chart axis (prevents axis inversion from
  hostile/huge values); test updated with non-negative assertions.
- redis: arm reconnect backoff on a command/read-timeout error (not just
  connect-fail) so a reachable-but-slow remote can't thrash
  connect+timeout each tick and stall the UI thread.
- redis: extract duplicated GET-into-buffer into redis_get_string.
- redis.h: document that the connect timeout does not bound getaddrinfo
  (recommend IP / /etc/hosts pin); full async resolve tracked in kwi #51.
- telemetry: reset SCAN cursor/building count on discover failure so a
  reconnect starts a clean pass.
- dev: rename dev_side_t.cpu_only -> applied_cpu_only; add
  side_mark_gap/side_set_status fan-out helpers; _Static_assert that the
  host selector list holds all discovered telemetry hosts.
- dev_graph: hoist duplicated 16384 into MB_AXIS_DEFAULT; drop dead
  dev_graph_destroy.
- docs: add telemetry env vars + dev:left|right keys to README; correct
  ROTATE_180 to a documented no-op (README + env.example).

Deferred follow-ups tracked as kwi #51 (DNS), #52 (waiting-for-data
state), #53 (opaque redis_client_t). Review artifact under .context/.

Host suites (dev_telemetry, dev_view, dev_hostlist) pass; cross-build clean.

+++ src/dev_telemetry.c
@@ -22,17 +22,20 @@ static float clamp_pct(const cJSON *o) {
     return (float)v;
 }
 
-/* Clamp a JSON number field to a non-negative MB count in [0, UINT32_MAX].
- * Guards the double->uint32 conversion against undefined behaviour on
- * negative, non-finite, or out-of-range values. */
+/* Clamp a JSON number field to a non-negative MB count in [0, INT32_MAX].
+ * Guards the double->integer conversion against undefined behaviour on
+ * negative, non-finite, or out-of-range values. The ceiling is INT32_MAX (not
+ * UINT32_MAX) so every value stays safe to cast to the int32_t the LVGL chart
+ * axis uses downstream — a larger value would wrap negative and invert the
+ * axis. Real RAM/VRAM totals are millions of MB at most, far below the cap. */
 static uint32_t clamp_mb(const cJSON *o) {
     if (!cJSON_IsNumber(o))
         return 0;
     double v = o->valuedouble;
     if (!isfinite(v) || v < 0.0)
         return 0;
-    if (v > (double)UINT32_MAX)
-        return UINT32_MAX;
+    if (v > (double)INT32_MAX)
+        return (uint32_t)INT32_MAX;
     return (uint32_t)v;
 }
